pyproject.toml
[project]
name = "langgraph-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Paul Williamson",email = "paulw.900430@gmail.com"}
]
readme = "readme.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph-cli[inmem] (>=0.2.10,<0.3.0)",
    "langgraph-sdk (>=0.1.69,<0.2.0)",
    "langchain-tavily (>=0.1.6,<0.2.0)",
    "langchain-openai (>=0.3.17,<0.4.0)",
    "yfinance (>=0.2.38,<0.3.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "bbc-news (>=1.0.2,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)"
]

[tool.poetry]
packages = [
    { include = "src/agent" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
